==> 06a_analyze_parentage_output_R.ipynb <==
# load depenedencies, get shape parameters

source('~/imports.R')
library(coda)
options(repr.plot.width=5, repr.plot.height=5)

dname = '~/eckertlab/brandon/teakettle/dependencies/'
cd(dname)
getrds()

usgibbs   <- readRDS('~/teakettle/data/gibbs/first/end_of_cyc_files/usgibbs_all_cycs.RDS')
upgibbs   <- readRDS('~/teakettle/data/gibbs/first/end_of_cyc_files/upgibbs_all_cycs.RDS')
ped.gibbs <- readRDS('~/teakettle/data/gibbs/first/end_of_cyc_files/pedgibbs_all_cycs.RDS')
ncyc      <- 150000

c(len(usgibbs),len(upgibbs),len(ped.gibbs))

# get plot posts so we include the full mapping area

pc <- matrix(rep(0,8),ncol=2)
xmin <- min(adult.loc[,1]) - 10
ymin = min(adult.loc[,2]) - 10
xmax = max(adult.loc[,1]) + 10
ymax = max(adult.loc[,2]) + 10
pc[4,1] <- pc[1,1] <- xmax
pc[2,2] <- pc[1,2] <- ymax
pc[3,1] <- pc[2,1] <- xmin
pc[4,2] <- pc[3,2] <- ymin
pcor <- pc

outpar <- matrix(NA,2,4) 
#burn  <- 3000                              #end of burn-in - original script had burn @ 3K, paper had 30K
burn   <- 30000                              #end of burn-in
thin   <- seq(burn,(length(usgibbs)),by=1500)  #thin the gibbs sequence
gmat   <- cbind(usgibbs[thin],upgibbs[thin])
outpar[,1]       <- apply(gmat,2,mean)
outpar[,2]       <- apply(gmat,2,sd)
outpar[,3:4]     <- t(apply(gmat,2,quantile,c(.025,.975)))
colnames(outpar) <- c('estimate','se','.025','.975')
rownames(outpar) <- c('us','up')
signif(outpar,2)

####  translate estimates, se, and quantiles to distances

outpardist = outpar
for (col in colnames(outpar)){
    outpardist[1,col] = mean.func(outpardist[1,col])
    outpardist[2,col] = mean.func(outpardist[2,col])
}
signif(outpardist,2)

# plot it
prob = function(d,u) {1 / (pi*u* ((1 + ((d**2)/u) )**2)  )}
us = 1700
up = 12000
usup = 2300
uslo = 1300
upup = 15000
uplo = 9200
usdf = data.frame(main = rep(NA,300), upper=rep(NA,300),lower=rep(NA,300),dist=seq(1,300))
updf = data.frame(main = rep(NA,300), upper=rep(NA,300),lower=rep(NA,300),dist=seq(1,300))
# us
for (i in 1:300){
    usdf[i,'main'] = prob(i,us)
    usdf[i,'upper']  = prob(i,usup)
    usdf[i,'lower']  = prob(i,uslo)

    updf[i,'main'] = prob(i,up)
    updf[i,'upper']  = prob(i,upup)
    updf[i,'lower']  = prob(i,uplo)
}
# pollen
plot(updf[,'dist'],updf[,'upper'],type='l',ylim=c(0,2.5e-04),lty=2,xlim=c(0,200*sqrt(2)),xlab="Distance (m)",ylab="Probability")
#par(new=T)
lines(updf[,'dist'],updf[,'lower'],type='l',ylim=c(0,2.5e-04),lty=2)
#par(new=T)
lines(updf[,'dist'],updf[,'main'],type='l',ylim=c(0,2.5e-04))
# seed
lines(usdf[,'dist'],usdf[,'upper'],type='l',ylim=c(0,2.5e-04),lty=2,col='red')
#par(new=T)
lines(usdf[,'dist'],usdf[,'lower'],type='l',ylim=c(0,2.5e-04),lty=2,col='red')
#par(new=T)
lines(usdf[,'dist'],usdf[,'main'],type='l',ylim=c(0,2.5e-04),col='red')
legend(150,0.0002,legend=c('Pollen','Seed'),col=c('black','red'),lty=c(1,1))

#### caclulate proportions below 150m

#pollen
probup = function(d) {1 / (pi*up* ((1 + ((d**2)/up) )**2)  )}
tot = integrate(probup,0,Inf)
frac = integrate(probup,0,150)
frac$value/tot$value

# seed
probus = function(d) {1 / (pi*us* ((1 + ((d**2)/us) )**2)  )}
stot = integrate(probus,0,Inf)
sfrac = integrate(probus,0,150)
sfrac$value/stot$value

# seed @ 50m
stot = integrate(probus,0,Inf)
sfrac = integrate(probus,0,50)
sfrac$value/stot$value

c(mean.func(mean(usgibbs)),mean.func(mean(usgibbs[thin])))

c(mean.func(mean(upgibbs)),mean.func(mean(upgibbs[thin])))

# plot shape parameters in R

par(mfcol=c(2,1))
plot(usgibbs,type='l',xaxt='n')
abline(h=mean(usgibbs[thin]))
axis(1,at=seq(5000,150000,by=5000))
plot(upgibbs,type='l',xaxt='n')
abline(h=mean(upgibbs[thin]))
axis(1,at=seq(5000,150000,by=5000))

effectiveSize(usgibbs[30000:len(usgibbs)])

effectiveSize(upgibbs[30000:len(upgibbs)])

len(upgibbs)

## Which parent pairs came up which percentage of the time?  

c(burn,ncyc,nadult,noff)

par.mat <- numeric(0)
# f.vec <- m.vec   <- matrix(0,noff,nadult+1) # never used
for(k in 1:noff)
{
    print(k)
    p.m <- matrix(0,nadult+1,nadult+1)
    for(d in burn:ncyc)
    {
        p.m[ped.gibbs[[d]][k,1],ped.gibbs[[d]][k,2]] <- 1 + p.m[ped.gibbs[[d]][k,1],ped.gibbs[[d]][k,2]] 
    }
    p.m <- p.m/sum(p.m)
    par.mat <- append(par.mat,list(p.m))
}

# Identify most frequent parent pair for each offspring.  M is most likely mother, f is most likely father, 
## per the percentage of time they were matched with offspring.
m <- rep(NA,noff)
f <- rep(NA,noff)
per <- rep(NA,noff) 
for(k in 1:noff){
      v <- matrix(0,nadult+1,nadult+1)
      v[which(par.mat[[k]][,]==max(par.mat[[k]]))]<-1

      m[k]<- which(apply(v,1,sum)==1)
      f[k]<- which(apply(v,2,sum)==1)
      per[k] <- max(par.mat[[k]])
}

table(m)

table(f)

# save RDS
DIR = '/home/lindb/teakettle/data/gibbs/first/end_of_cyc_files'
objects <- c('par.mat','m','f','per','outpar')

for (obj in objects){
    filename = sprintf("%s.RDS",obj)
    cat(sprintf('saving %s\n',filename))
    saveRDS(get(obj),fname(DIR,filename))
}

DIR = '/home/lindb/teakettle/data/gibbs/first/end_of_cyc_files'
cd(DIR)
objects <- c('par.mat','m','f','per','outpar')

for (obj in objects){
    getrds(,pattern=obj)
}

## Plot parent-offspring connections

### determine connections
x0 <- numeric(0)
y0 <- x0
x1 <- x0
y1 <- x0
x0b <- x0
y0b <- x0
x1b <- x0
y1b <- x0
x0c <- x0
y0c <- x0
x1c <- x0
y1c <- x0
x0.o <- x0
y0.o <- x0
x1.o <- x0
y1.o <- x0
x0b.o <- x0
y0b.o <- x0
x1b.o <- x0
y1b.o <- x0
xo <- x0
yo <- x0
xf <- x0
yf <- x0

for (k in 1:noff){
    if(m[k]<nadult+1){      # if mother is in the plot
        # offspring locations
        x0 <- c(x0,off.loc[k,1])
        y0 <- c(y0,off.loc[k,2])
        # mother locations
        x1 <- c(x1,adult.loc[m[k],1])
        y1 <- c(y1,adult.loc[m[k],2])
        
        if(f[k]<nadult+1) { # if father is in the plot
            # mother locations
            x0b <- c(x0b,adult.loc[m[k],1])
            y0b <- c(y0b,adult.loc[m[k],2])
            # father locations
            x1b <- c(x1b,adult.loc[f[k],1])
            y1b <- c(y1b,adult.loc[f[k],2])
            x0c <- c(x0c,adult.loc[f[k],1]) # redundant for some reason
            y0c <- c(y0c,adult.loc[f[k],2]) # redundant for some reason
            # offspring locations
#             x1c <- c(x0,off.loc[k,1]) # this can't be right
#             y1c <- c(y0,off.loc[k,2]) # this can't be right
            x1c <- c(x1c,off.loc[k,1])  # replace with this
            y1c <- c(y1c,off.loc[k,2])  # replace with this
        } 
        if(f[k]==nadult+1){ # if father is outside the plot
            # mother locations
            x0b.o <- c(x0b.o,adult.loc[m[k],1])
            y0b.o <- c(y0b.o,adult.loc[m[k],2])
            # out-of-plot father locations
            md <- which(d.edge.a[m[k],]==min(d.edge.a[m[k],]))
            x1b.o[k] <- edgemid[md,1]
            y1b.o[k] <- edgemid[md,2] 
        }
    }
    if(m[k]==nadult+1){     # if mother is outside the plot
        # offspring locations
        x0.o <- c(x0.o,off.loc[k,1])
        y0.o <- c(y0.o,off.loc[k,2])
        md <- which(d.edge.o[k,]==min(d.edge.o[k,]))
        # out-of-plot mother locations
        x1.o <- c(x1.o,edgemid[md,1])
        y1.o <- c(y1.o,edgemid[md,2])
        # don't need combos of in- or out-of-plot fathers (otherwise need if statments nested here)
    }
    # added these so that I could get off-father distances without condition of mother inside plot
    if (f[k]<nadult+1){
        # offspring locations
        xo <- c(xo,off.loc[k,1]) 
        yo <- c(yo,off.loc[k,2])
        # father locations
        xf <- c(xf,adult.loc[f[k],1])
        yf <- c(yf,adult.loc[f[k],2])
    }
}

### plot connections
#cornerposts  
X <- pcor[,1]
Y <- pcor[,2]
plot(X,Y,main='father to offspring')
segments(xo,yo,xf,yf,lty=3,lwd =1,col='grey') #inplot father to offspring
#adults
#d2 <- w # viewing with symbol size representative of weight is messy
d2 <- rep(4,nadult)
symbols(adult.loc[,1],adult.loc[,2],squares=d2,inches=F,fg=4,add=T)
#seedlings
d4 <- rep(4,noff)
symbols(off.loc[,1],off.loc[,2],circles=d4,inches=F,fg=3,add=T)

#segments(x0c,y0c,x1c,y1c,lty=2,lwd =1) #old (only includes o-f pairs conditional on known mother)
#segments(xo,yo,xf,yf,lty=3,lwd =1,col='grey') #inplot father to offspring

#cornerposts  
plot(X,Y,main='mother to offspring')
symbols(adult.loc[,1],adult.loc[,2],squares=d2,inches=F,fg=4,add=T)
symbols(off.loc[,1],off.loc[,2],circles=d4,inches=F,fg=3,add=T)

segments(x0,y0,x1,y1,lwd=1)  #inplot mother to offspring

#cornerposts  
plot(X,Y,main='father to mother')
symbols(adult.loc[,1],adult.loc[,2],squares=d2,inches=F,fg=4,add=T)
symbols(off.loc[,1],off.loc[,2],circles=d4,inches=F,fg=3,add=T)

segments(x0b,y0b,x1b,y1b,col="red",lwd=1) #inplot father to mother

c(len(x1b),len(y1b),len(x1c),len(y1c))

## hist of mother-off-father distances

distance = function(a0,a1,b0,b1){
    ddd <- rep(0,len(a0))
    for (i in 1:len(a0)){
        adiff = (a0[i]-a1[i])**2
        bdiff = (b0[i]-b1[i])**2
        diff = sqrt(adiff+bdiff)
        ddd[i] <- diff
    }
    return (ddd)
}

#### mother-offspring

#x0 y0 x1 y1
m.off <- distance(x0,x1,y0,y1)

lapply(list(x0,x1,y0,y1),FUN=len)

c(mean(m.off),median(m.off))

mean.func(mean(usgibbs))

mean.func(mean(usgibbs[thin]))

hist(m.off, main="mother-offspring distances",xlab ="meters")

#### father-offspring

#f.off <- distance(x0c,x1c,y0c,y1c)
f.off <- distance(xo,xf,yo,yf)

#lapply(list(x0c,x1c,y0c,y1c),FUN=len)
lapply(list(xo,xf,yo,yf),FUN=len)

hist(f.off,main="father-offspring distances",xlab="meters",breaks=25)

#### mother-father

lapply(list(x0b,y0b,x1b,y1b),FUN=len)

m.f <- distance(x0b,x1b,y0b,y1b)

c(mean(m.f),median(m.f))

len(m.f)

hist(m.f,main="mother-father distances",xlab="meters",breaks=16)

# rep 2

usgibbs2 <- readRDS('~/teakettle/data/gibbs/first_rep2/end_of_cyc_files/usgibbs_all_cycs.RDS')
upgibbs2 <- readRDS('~/teakettle/data/gibbs/first_rep2/end_of_cyc_files/upgibbs_all_cycs.RDS')
#ped.gibbs <- readRDS('~/teakettle/data/gibbs/first_rep2/end_of_cyc_files/pedgibbs_all_cycs.RDS')

par(mfcol=c(2,1))
plot(usgibbs2,type='l')
abline(h=mean(usgibbs2[thin]))
plot(upgibbs2,type='l')
abline(h=mean(upgibbs2[thin]))

c(mean(usgibbs[thin]),mean(usgibbs2[thin]))

c(mean.func(mean(usgibbs[thin])),mean.func(mean(usgibbs2[thin])))

len(usgibbs2)

effectiveSize(usgibbs2[30000:len(usgibbs2)])

effectiveSize(upgibbs2)

# rep 3

usgibbs3 <- readRDS('~/teakettle/data/gibbs/first_rep3/end_of_cyc_files/usgibbs_all_cycs.RDS')
upgibbs3 <- readRDS('~/teakettle/data/gibbs/first_rep3/end_of_cyc_files/upgibbs_all_cycs.RDS')
#ped.gibbs <- readRDS('~/teakettle/data/gibbs/first_rep3/end_of_cyc_files/pedgibbs_all_cycs.RDS')

par(mfcol=c(2,1))
plot(usgibbs3,type='l')
abline(h=mean(usgibbs3[thin]))
plot(upgibbs3,type='l')
abline(h=mean(upgibbs3[thin]))

c(mean(usgibbs[thin]),mean(usgibbs2[thin]),mean(usgibbs3[thin]))

c(mean.func(mean(usgibbs[thin])),mean.func(mean(usgibbs2[thin])),mean.func(mean(usgibbs3[thin])))

effectiveSize(usgibbs3[30000:len(usgibbs3)])

effectiveSize(upgibbs3[30000:len(upgibbs3)])

effectiveSize(upgibbs3)
