==> 11_moran-and-clark_output_analysis_py.ipynb <==
imp = open('/home/lindb/pythonimports.py').read()
exec(imp)
from __future__ import division
%matplotlib inline 

##### R magic voodoo

### set up R
def setup_r(): 
    os.environ['R_HOME'] = '/home/lindb/g/R3/lib64/R/' 
    os.environ['LD_LIBRARY_PATH'] = "%s/lib:%s" % (os.environ['R_HOME'], 
                                                   os.environ['LD_LIBRARY_PATH'])
setup_r()
import readline
import rpy2.robjects as robjects
from rpy2.robjects import pandas2ri as pd2ri
pd2ri.activate()
r = robjects.r
%reload_ext autoreload
%autoreload 2
%matplotlib inline
%load_ext rpy2.ipython
%matplotlib inline

%%R
source('/home/lindb/imports.R')
sessionInfo()

def get_r_series(key):
    s = pd.Series(get_r(key))
    s.index = get_r("names(%s)" % key)
    return s

def get_r_df(key):
    df = pd.DataFrame(get_r(key))
    try:
        rname = get_r("rownames(%s)" % key)
        df.index = rname
    except:
        pass
    
    try:
        cname = get_r("colnames(%s)" % key)
        df.columns = cname
    except:
        pass
    return df

def get_r(key):
    return r(key)

##### dependencies

pg = pd.read_csv('/home/lindb/teakettle/postmoran/moran_analysis/assignments.txt',index_col=0,sep='\t')
pg.head()

geno = pd.read_csv('/home/lindb/teakettle/data/moran/moran_infile_final_jan_wconecounts.txt',header=0,sep='\t')
geno = geno[geno.columns[:6]]
geno = geno.loc[np.arange(0,nrow(geno),2),:]
geno.index = geno['samp'].tolist()
geno = geno[geno.columns[1:]]
geno.head()

# combine data
comb = pd.merge(pg,geno,left_index=True,right_index=True)
comb.head()

dbh = {}
for samp in comb.index:
    dbh[samp] = comb['dbh_11'][samp]

comb.to_csv('/home/lindb/teakettle/postmoran/moran_analysis/infile_gts_combined.txt',header=True,sep='\t')

comb = pd.read_csv('/home/lindb/teakettle/postmoran/moran_analysis/infile_gts_combined.txt',header=0,index_col=0,sep='\t')
comb.head()

comb.loc['uc3_new076',['mom','dad']]

%%R
# read into R
par.mat <- readRDS('/home/lindb/teakettle/data/gibbs/first/end_of_cyc_files/par.mat.RDS')
bframe1 = readRDS('/home/lindb/teakettle/data/moran/dependencies/bframe1.RDS')
bframe2 = readRDS('/home/lindb/teakettle/data/moran/dependencies/bframe2.RDS')
adults <- bframe1[seq(1,nrow(bframe1),2),'samp']
seedlings <- bframe2[seq(1,nrow(bframe2),2),'samp']

# look at fractional parentage using assignment probabilities from moran & clark cycles

DIR = '/home/lindb/teakettle/postmoran/moran_analysis/seedout'
if not op.exists(DIR):
    mkdir(DIR)

%%R
# get all potential parents
DIR = '/home/lindb/teakettle/postmoran/moran_analysis/seedout'
for (k in 1:len(par.mat)){
    # get df with all non-zero probs
    whiches = which(par.mat[[k]] != 0,arr.ind=T)
    df = data.frame(whiches)
    colnames(df) = c('mom','dad') # mom and dad columns
    stopifnot(round(sum(par.mat[[k]][whiches]),2) == 1)
    df[,'prob'] = par.mat[[k]][whiches] # add probs
    # save df
    sname = seedlings[k]
    write.table(df,fname(DIR,sprintf("%s.txt",sname)),sep='\t',row.names=F)
}

# translate num to samp name
As = get_r_df('adults')
As.loc[599] = 'out-of-plot'
adults = {}
for row in As.index:
    adults[row+1] = As[0][row]
Ss = get_r_df('seedlings')
seeds = {}
for row in Ss.index:
    seeds[row+1] = Ss[0][row]

# for each seedling get a df with potential parents with probabilities
DIR = '/home/lindb/teakettle/postmoran/moran_analysis/seedout'
pars = {}
for f in fs(DIR):
    samp = op.basename(f).split(".txt")[0]
    df = pd.read_csv(f,sep='\t')
    pars[samp] = df

pkldump(pars,'/home/lindb/teakettle/postmoran/moran_analysis/pars.pkl')

pkldump(adults,'/home/lindb/teakettle/postmoran/moran_analysis/adults.pkl')

pars = pickle.load(open('/home/lindb/teakettle/postmoran/moran_analysis/pars.pkl'))

adults = pickle.load(open('/home/lindb/teakettle/postmoran/moran_analysis/adults.pkl'))

pars[samp].head(),sum(pars[samp]['prob'])

# look at distribution of max probabilities
maxs = []
for samp in pars.keys():
    maxs.append(pars[samp]['prob'].max())
plt.hist(maxs)

# how does the probabilities shake out over probability classes?
sums = len(maxs)
for thresh in np.arange(0,1,.1):
    l = len([x for x in maxs if x > thresh])
    sums +- l
    print thresh,l
sums

# replicate moran's method of IDing most likely father/mother pair
mothers = []
fathers = []
for samp in pars:
    row,col,probidx = pars[samp].idxmax()
    row,col,prob = pars[samp].loc[probidx,:]
    mothers.append(row)
    fathers.append(col)

# check and compare to table(m) from 06a...ipynb
t = table(mothers)
t

# check and compare using samp names
# print mother samp name, noff, mother dbh
for k,v in t.iteritems():
    if adults[int(k)] == 'out-of-plot':
        print '\t',adults[int(k)],v
    else:
        print adults[int(k)],v, dbh[adults[int(k)]]

# check and compare to table(f) from 06a...ipynb
t2 = table(fathers)
t2

# print father samp name, noff, father dbh
for k,v in t2.iteritems():
    if adults[int(k)] == 'out-of-plot':
        print '\t',adults[int(k)],v
    else:
        print adults[int(k)],v, dbh[adults[int(k)]]

# make sure each df['prob'] sums to 1.0
for samp in pars:
    assert "%.2f" % sum(pars[samp]['prob'].tolist()) == '1.00'

pars[samp].head()

# create blank dataframe
momheat = OrderedDict()
dadheat = OrderedDict()
for adult in adults.values():
    momheat[adult] = OrderedDict()
    dadheat[adult] = OrderedDict()
    for juv in seeds.values():
        momheat[adult][juv] = float(0.0)
        dadheat[adult][juv] = float(0.0)
# fill in non-zero probability values
for samp in pars:
    for idx in pars[samp].index:
        mom,dad,prob = pars[samp].loc[idx,]
        mom = adults[int(mom)]
        dad = adults[int(dad)]
        if momheat[mom][samp] == 0.0:
            momheat[mom][samp] = float(prob)
        else:
            momheat[mom][samp] += float(prob)
        if dadheat[dad][samp] == 0.0:
            dadheat[dad][samp] = float(prob)
        else:
            dadheat[dad][samp] += float(prob)

DIR = '/home/lindb/teakettle/postmoran/moran_analysis/heat'
if not op.exists(DIR):
    mkdir(DIR)
mh = pd.DataFrame(momheat)
mh.to_csv(op.join(DIR,'momheat.txt'),sep='\t')
dh = pd.DataFrame(dadheat)
dh.to_csv(op.join(DIR,'dadheat.txt'),sep='\t')

mh.head()

head(dh)

# assert probs sum to 1
for row in dh.index:
    assert round(sum(dh.loc[row,:].tolist()),2) == 1

# assert probs sum to 1
for row in mh.index:
    assert round(sum(mh.loc[row,:].tolist()),2) == 1

%%R
# prep matrices for heatmap
mom = read.csv('~/teakettle/postmoran/moran_analysis/heat/momheat.txt',sep='\t',stringsAsFactors=F)
rownames(mom) <- mom$X
seedlings <- rownames(mom)
adults <- colnames(mom)
mom <- mom[,-1]
dad = read.csv('/home/lindb/teakettle/postmoran/moran_analysis/heat/dadheat.txt',sep='\t',stringsAsFactors=F)
rownames(dad) <- dad$X
dad <- dad[,-1]
m = as.matrix(mom)
d = as.matrix(dad)
# log transform non-zero values
whiches    = which(m != 0)
m[whiches] = log(m[whiches],base=10) #,base=10
whiches    = which(d != 0)
d[whiches] = log(d[whiches],base=10)

%%R
# determine abline locations
# which seedlings belong to which treatments
bcs = which(grepl('bc',seedlings) == T)
ucs = which(grepl('uc',seedlings) == T)
uns = which(grepl('un',seedlings) == T)
bcseedline = max(bcs) # border line
ucseedline = max(ucs) # border line
# which adults belong to which treatments
abc = which(grepl('bc',adults) == T)
auc = which(grepl('uc',adults) == T)
aun = which(grepl('un',adults) == T)
bcad = max(abc)
ucad = max(auc)
# was going to use this to label margins, but will just do this in Illustrator
bcsat = (diff(range(bcs))/2)+min(bcs)
ucsat = (diff(range(ucs))/2)+min(ucs)
unsat = (diff(range(uns))/2)+min(uns)
abcat = (diff(range(abc))/2)+min(abc)
aucat = (diff(range(auc))/2)+min(auc)
aunat = (diff(range(aun))/2)+min(aun)

%%R
# plot heatmap for fractional parentage for seedling assignment to a given mother tree
#pdf('/home/lindb/teakettle/postmoran/figs/momheat.pdf')
my_palette <- colorRampPalette(rev(c("black","red","orange","yellow","white")))(n = 499)
heatmap(m,col=my_palette,Rowv=NA,Colv=NA,scale="none",
        labRow=FALSE,labCol=FALSE,xlab="Mothers",ylab="Seedlings",margins=c(2,2),
        add.expr = abline(h=c(bcseedline,ucseedline), v=c(bcad,ucad),lwd=.75,col="white")
       )
#dev.off()

%%R
# plot heatmap for fractional parentage for seedling assignment to a given father tree
#pdf('/home/lindb/teakettle/postmoran/figs/dadheat.pdf')
heatmap(d,col=my_palette,Rowv=NA,Colv=NA,scale="none",
        labRow=FALSE,labCol=FALSE,xlab="Fathers",ylab="Seedlings",margins=c(4,4),
        add.expr = abline(h=c(bcseedline,ucseedline), v=c(bcad,ucad),lwd=.75,col="white")
       )
mtext(c('bc','uc','un'),side=2,at=c(bcsat,ucsat,unsat))
#dev.off()

%%R
# make sure partitions are where I expect them to be
lapply(list(bcs,ucs,uns),FUN=len)

%%R
# make sure partitions are where I expect them to be
lapply(list(abc,auc,aun),FUN=len)

# triple check blocking of figure

# ideas to test

* compare all below by treatment and between seeds/pollen (mom to seed, father to seed, father to mother)


1. mean dispersal distance within treatment 
    1. father to seed could be an 'extended/effective dispersal'
1. mean dispersal distance out of treatment - but how to account for proximity of other plots?
    1. can address concern with permutations
1. number of out of plot dispersals 
    1. may be able to account for prox of plots by lumping 'unknown parents' and 'out-of-plot' parents
1. ratio of within to without 
    1. number of dispersals
    1. distances of dispersals


* using vector of potential parents (PP)


1. mean number of PPs within treatment
1. mean number of PPs outside of treatment
1. also by weighting PP by prob

### dependencies

# transpose dataframes so I can more easily access fractional parentage by key=seedling.ID
dhtrans = dh.transpose()
mhtrans = mh.transpose()

dhtrans.head()

comb.head()

# get utm coordinates for each sample
utm = {}
for samp in comb.index.tolist():
    utm[samp] = [comb['xutm'][samp],
                 comb['yutm'][samp]]

def collapsedf(df,par):
    # sum probs for a given mom or dad if they appear in more than one pedigree
    # don't actually need this, but oh well
    if nrow(df) == 0:
        return df
    df2 = OrderedDict()
    for x in ['mom','dad','prob']:
        df2[x] = OrderedDict()
    if par == 'mom':
        col = df.columns[0]
        seedcol = df.columns[1]
    else:
        col = df.columns[1]
        seedcol = df.columns[0]
    assert luni(df[seedcol]) == 1
    seed = uni(df[seedcol])[0]
    if luni(df[col]) == nrow(df):
        return df
    else:
        count = 0
        df2[col] = OrderedDict()
        df2[seedcol] = OrderedDict()
        df2['prob'] = OrderedDict()
        for p in uni(df[col]):
            d = df[df[col]==p].copy()
            sums = sum(d['prob'])
            df2[col][count] = p
            df2[seedcol][count] = seed
            df2['prob'][count] = sums
            count += 1
    return pd.DataFrame(df2)
def distance(i1 ,i2):
    # euclidean distance
    xdiff = (utm[i1][0] - utm[i2][0])**2
    ydiff = (utm[i1][1] - utm[i2][1])**2
    return np.sqrt(xdiff+ydiff)
def wtd_dist(seed,par,bytx=False,without=False,byplot=False,bynum=False,weights=False):
    # weight par-off or mother-father distances by probability of parentage
    df = pars[seed].copy()
    if par == 'mom':
        df = df[df['mom'] != 600].copy()
        df.loc[:,'mom'] = [adults[int(x)] for x in df['mom']]
        df.loc[:,'dad'] = [seed for x in range(len(df))]
        df = collapsedf(df,'mom')
    elif par == 'dad':
        df = df[df['dad'] != 600].copy()
        df.loc[:,'dad'] = [adults[int(x)] for x in df['dad']]
        df.loc[:,'mom'] = [seed for x in range(len(df))]
        df = collapsedf(df,'dad')
    else:
        df = df[df['mom'] != 600].copy()
        df = df[df['dad'] != 600].copy()
        df.loc[:,'mom'] = [adults[int(x)] for x in df['mom'].tolist()]
        df.loc[:,'dad'] = [adults[int(x)] for x in df['dad'].tolist()] 
#        return df
    if bytx == True or byplot == True:
        if bytx:
            tx = seed[:2]
            Dict = txsamps
        if byplot:
            tx = seed[:3]
            Dict = plotsamps
#         if luni(df['mom']) == 1: # but what about nrow(df) == 1
        if par == 'mom':
            cols = ['mom']
#         elif luni(df['dad']) == 1:
        elif par == 'dad':
            cols = ['dad']
        else:
            cols = ['mom','dad']
        for col in cols:
            if without == False: # if only looking at samps within the tx/plot
                df = df[df[col].isin(Dict[tx])].copy()
            else:
                df = df[df[col].isin(Dict[tx])==False].copy()
    if bynum == True:
        if weights == False:
            return(nrow(df))
        else:
            probsum = sum(df['prob'])
            return [nrow(df),probsum]
    if nrow(df) == 0:
        return 'NA'
    sums = 0
    probsum = sum(df['prob'])
    df.index = range(nrow(df))
    for idx in df.index:
        ind1,ind2,prob = df.loc[idx,]
        d = distance(ind1,ind2)
        p = prob/probsum
        s = d*p
        sums += s
    return sums

# overal mean dispersal from moran & clark outputs

# ID most likely father/mother pair for each seedling
assignment = {}
for samp in pars:
    row,col,probidx = pars[samp].idxmax()
    row,col,prob = pars[samp].loc[probidx,:]
    assignment[samp] = [adults[int(row)],adults[int(col)]]
#    print samp,assignment[samp]

#### mother-offspring

moffdisp = {}
moffdisptx = {}
for samp in assignment:
    plot = samp[:3]
    tx   = samp[:2]
    if not plot in moffdisp:
        moffdisp[plot] = []
    if not tx in moffdisptx:
        moffdisptx[tx] = []
    mom,dad = assignment[samp]
    if mom != 'out-of-plot':
        d = distance(mom,samp)
        moffdisp[plot].append(d)
        moffdisptx[tx].append(d)
for tx in moffdisptx:
    print tx,np.mean(moffdisptx[tx]),np.median(moffdisptx[tx])

print stats.kruskal(moffdisptx['bc'],moffdisptx['uc'],moffdisptx['un']).pvalue
for i,txi in enumerate(moffdisptx.keys()):
    for j,txj in enumerate(moffdisptx.keys()):
        if i > j:
            print txi,round(np.median(moffdisptx[txi]),2),txj,round(np.median(moffdisptx[txj]),2)
            print stats.kruskal(moffdisptx[txi],moffdisptx[txj])

type(range(1,4,1))

#yerr = [np.std(moffdisptx[tx]) for tx in moffdisptx]
f = plt.figure()
#plt.bar(range(3),[np.median(moffdisptx[tx]) for tx in moffdisptx],yerr=yerr)[-1]
plt.boxplot([moffdisptx[tx] for tx in moffdisptx])
plt.xticks(range(1,4,1),list(['UN','UC','BC']))
plt.ylabel('Mother-Offspring Distances (m)')
plt.ylim((-1,145))
f.savefig('/home/lindb/teakettle/postmoran/figs/moff_true_parents.pdf')

#### mother-father

mfdisp = {}
mfdisptx = {}
for samp in assignment:
    plot = samp[:3]
    tx = samp[:2]
    if not plot in mfdisp:
        mfdisp[plot] = []
    if not tx in mfdisptx:
        mfdisptx[tx] = []
    mom,dad = assignment[samp]
    if not 'out-of-plot' in [mom,dad]:
        d = distance(mom,dad)
        mfdisp[plot].append(d)
        mfdisptx[tx].append(d)
for tx in mfdisptx:
    print tx,np.mean(mfdisptx[tx]),np.median(mfdisptx[tx])

print stats.kruskal(mfdisptx['bc'],mfdisptx['uc'],mfdisptx['un']).pvalue
for i,txi in enumerate(mfdisptx.keys()):
    for j,txj in enumerate(mfdisptx.keys()):
        if i > j:
            print txi,txj
            print stats.kruskal(mfdisptx[txi],mfdisptx[txj])

#yerr = [np.std(moffdisptx[tx]) for tx in moffdisptx]
f = plt.figure()
#plt.bar(range(3),[np.median(moffdisptx[tx]) for tx in moffdisptx],yerr=yerr)[-1]
plt.boxplot([mfdisptx[tx] for tx in mfdisptx])
plt.xticks(range(1,4,1),list(['UN','UC','BC']))
plt.ylabel('Father-Mother Distances (m)')
plt.ylim((-1,225))
f.savefig('/home/lindb/teakettle/postmoran/figs/mf_true_parents.pdf')

#### father-offspring

doffdisp = {}
doffdisptx = {}
for samp in assignment:
    plot = samp[:3]
    tx = samp[:2]
    if not plot in doffdisp:
        doffdisp[plot] = []
    if not tx in doffdisptx:
        doffdisptx[tx] = []
    mom,dad = assignment[samp]
    if not dad == 'out-of-plot':
        d = distance(samp,dad)
        doffdisp[plot].append(d)
        doffdisptx[tx].append(d)
for tx in doffdisptx:
    print tx,np.mean(doffdisptx[tx]),np.median(doffdisptx[tx])

print stats.kruskal(doffdisptx['bc'],doffdisptx['uc'],doffdisptx['un']).pvalue
for i,txi in enumerate(doffdisptx.keys()):
    for j,txj in enumerate(doffdisptx.keys()):
        if i > j:
            print txi,txj
            print stats.kruskal(doffdisptx[txi],doffdisptx[txj]).pvalue

# overall mean dispersal from fractional parentage

## $D_{i} = \sum_{j=1}^n (d_{j}*p_{j}) / n$
### $p_{j} = $ 

where $D_i$ = weighted mean distance for $seedling_i$ over each *j* of the *n* non-zero parent-offspring or mother-father probabilities each with distance $d_j$ and probability $p_j$

# problem is that I can only calc distances for non out-of-plot, 
# therefore it is likely that for any list of possible in-plot parents sum(probs) << 1

### mother-offspring distances
# can't do for each mother since vector of off probs won't sum to 1
# can only do for each seed
moff = {}
for seed in mh.index.tolist():
    moff[seed] = wtd_dist(seed,'mom')
#     print seed,moff[seed]

vals = []
[vals.append(val) for val in moffdisptx[tx] for tx in moffdisptx]
print np.mean(vals),np.median(vals)
plt.hist(vals,bins=100)[-1]

vals = [val for val in moff.values() if not val == 'NA']
print np.mean(vals),np.median(vals)
plt.hist(vals,bins=100)[-1]

### father-offspring distances
# can't do for each father since vector of off probs won't sum to 1
# can only do for each seed
doff = {}
for seed in mh.index.tolist():
    doff[seed] = wtd_dist(seed,'dad')
#     print seed,doff[seed]

print np.mean(doff.values()),np.median(doff.values())
plt.hist(doff.values(),bins=15)[-1]

# mother-father distances
mfs = []
for seed in pars:
    mfs.append( wtd_dist(seed,'par'))

plt.hist([x for x in mfs if x!='NA'],bins=50)[-1]

# mean dispersal by tx using fractional parentage

# list of samps by tx
txsamps = {}
txseeds = {}
for pop in uni(comb['pop'].tolist()):
    if not pop == 'us':
        df = comb[comb['pop']==pop].copy()
        txsamps[pop] = df.index.tolist()
        txseeds[pop] = [x for x in df.index.tolist() if 'new' in x]
        print pop,len(txsamps[pop]),len(txseeds[pop])

# list of samps by plot
plotsamps  = {}
plotseeds  = {}
plotadults = {}
for samp in comb.index.tolist():
    plot = samp[:3]
    if not plot == 'us3':
        if not plot in plotsamps:
            plotsamps[plot] = []
        if not plot in plotseeds:
            plotseeds[plot] = []
        if not plot in plotadults:
            plotadults[plot] = []
        if 'new' in samp:
            plotseeds[plot].append(samp)
        else:
            plotadults[plot].append(samp)
        plotsamps[plot].append(samp) # want adults and seeds
        assert len(plotsamps[plot]) == len(plotseeds[plot]) + len(plotadults[plot])
for plot in plotsamps:
    print plot, len(plotsamps[plot]),len(plotseeds[plot]),len(plotadults[plot])

'new' in 'bc1_1179','new' in 'bc1_new003'

#### mother-offspring distances by treatment

txmoff = {}
for tx in txseeds:
    txmoff[tx] = {}
    for samp in txseeds[tx]:
        txmoff[tx][samp] = wtd_dist(samp,'mom')

f = plt.figure()
vals = {}
for tx in txsamps:
    vals[tx] = [val for val in txmoff[tx].values() if not val == 'NA']
    print tx,np.mean(vals[tx]),np.median(vals[tx])
plt.boxplot([vals[tx] for tx in vals])
f.set_size_inches(6, 8)
plt.xticks(range(1,4,1),list([tx.upper() for tx in vals]))
plt.ylim((0,535))
plt.ylabel('Mother-Offspring Distances (m)')
f.savefig('/home/lindb/teakettle/postmoran/figs/txmoff.pdf')

f = plt.figure()
vals = {}
for tx in txsamps:
    v = [val for val in txmoff[tx].values() if not val == 'NA']
    vals[tx] = np.median(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list([v.upper() for v in vals.keys()]))
plt.ylabel('Distance (m)')
plt.ylim((0,60))
#plt.title('Mother-offspring distances by treatment')
f.savefig('/home/lindb/teakettle/postmoran/figs/txmoffbar.pdf')

bcs = [x for x in txmoff['bc'].values() if not x == 'NA' ]
ucs = [x for x in txmoff['uc'].values() if not x == 'NA' ]
uns = [x for x in txmoff['un'].values() if not x == 'NA' ]
bcucuns = [bcs,ucs,uns]
print stats.kruskal(bcs,ucs,uns)
for i,txi in enumerate(['bc','uc','un']):
    for j,txj in enumerate(['bc','uc','un']):
        if i > j:
            print txi,round(np.median(bcucuns[i]),2),txj,round(np.median(bcucuns[j]),2)
            print stats.kruskal(bcucuns[i],bcucuns[j])

#### father-offspring distances by treatment

txdoff = {}
for tx in txsamps:
    txdoff[tx] = {}
    for samp in txseeds[tx]:
        txdoff[tx][samp] = wtd_dist(samp,'dad')

vals = {}
for tx in txsamps:
    v = [val for val in txdoff[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Father-offspring distances by treatment')

#### mother-father distances by treatment

txmf = {}
for tx in txsamps:
    txmf[tx] = {}
    for samp in txseeds[tx]:
        txmf[tx][samp] = wtd_dist(samp,'par')

[v.upper() for v in vals.keys()]

f = plt.figure()
vals = {}
for tx in txmf:
    vals[tx] = [val for val in txmf[tx].values() if not val == 'NA']
    print tx,np.mean(vals[tx]),np.median(vals[tx])
plt.boxplot([vals[tx] for tx in vals])
f.set_size_inches(6, 8)
plt.xticks(range(1,4,1),list([tx.upper() for tx in vals]))
plt.ylabel('Father-Mother Distances (m)')
f.savefig('/home/lindb/teakettle/postmoran/figs/txmf.pdf')

f = plt.figure()
vals = {}
for tx in txsamps:
    v = [val for val in txmf[tx].values() if not val == 'NA']
    vals[tx] = np.median(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list([v.upper() for v in vals.keys()]))
plt.ylim(0,155)
plt.ylabel('Distance (m)')
#plt.title('Mother-father distances by treatment')
f.savefig('/home/lindb/teakettle/postmoran/figs/txmfbar.pdf')

bcs = [x for x in txmf['bc'].values() if not x == 'NA' ]
ucs = [x for x in txmf['uc'].values() if not x == 'NA' ]
uns = [x for x in txmf['un'].values() if not x == 'NA' ]
bcucuns = [bcs,ucs,uns]
print stats.kruskal(bcs,ucs,uns)
for i,txi in enumerate(['bc','uc','un']):
    for j,txj in enumerate(['bc','uc','un']):
        if i > j:
            print txi,round(np.median(bcucuns[i]),2),txj,round(np.median(bcucuns[j]),2)
            print stats.kruskal(bcucuns[i],bcucuns[j])

# mean dispersal within tx using fractional parentage

### mean dispersal by tx to within tx using fractional parentage
#### mother-offspring distances

wtxmoff = {}
for tx in txseeds:
    wtxmoff[tx] = {}
    for samp in txseeds[tx]:
        wtxmoff[tx][samp] = wtd_dist(samp,'mom',bytx=True)

vals = {}
for tx in txsamps:
    v = [val for val in wtxmoff[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-offspring distances within treatment')

#### father-offspring distances

wtxdoff ={}
for tx in txseeds:
    wtxdoff[tx] = {}
    for samp in txseeds[tx]:
        wtxdoff[tx][samp] = wtd_dist(samp,'dad',bytx=True)

vals = {}
for tx in txsamps:
    v = [val for val in wtxdoff[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('father-offspring distances within treatment')

#### mother-father distances

wtxmf = {}
for tx in txseeds:
    wtxmf[tx] = {}
    for samp in txseeds[tx]:
        wtxmf[tx][samp] = wtd_dist(samp,'par',bytx=True)

vals = {}
for tx in txsamps:
    v = [val for val in wtxmf[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-father distances within treatment')

### mean dispersal by tx to outside of tx using fractional parentage

#### mother-offspring distances

otxmoff = {}
for tx in txseeds:
    otxmoff[tx] = {}
    for samp in txseeds[tx]:
        otxmoff[tx][samp] = wtd_dist(samp,'mom',bytx=True,without=True)

vals = {}
for tx in txsamps:
    v = [val for val in otxmoff[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-offspring distances outside of treatment')

#### father-offspring distances

otxdoff = {}
for tx in txseeds:
    otxdoff[tx] = {}
    for samp in txseeds[tx]:
        otxdoff[tx][samp] = wtd_dist(samp,'dad',bytx=True,without=True)

vals = {}
for tx in txsamps:
    v = [val for val in otxdoff[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Father-offspring distances outside of treatment')

#### mother-father distances

otxmf = {}
for tx in txseeds:
    otxmf[tx] = {}
    for samp in txseeds[tx]:
        otxmf[tx][samp] = wtd_dist(samp,'par',bytx=True,without=True)

vals = {}
for tx in txsamps:
    v = [val for val in otxmf[tx].values() if not val == 'NA']
    vals[tx] = np.mean(v)
    print tx,np.mean(v),np.median(v)
x = np.arange(3)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-father distances outside of treatment')

# mean dispersal within plot using fractional parentage

#### mother-offspring distances

wplotmoff = {}
for plot in plotsamps:
    wplotmoff[plot] = {}
    for samp in plotseeds[plot]:
        wplotmoff[plot][samp] = wtd_dist(samp,'mom',byplot=True)

vals = {}
for plot in plotsamps:
    v = [val for val in wplotmoff[plot].values() if not val == 'NA']
    vals[plot] = np.mean(v)
    print plot,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-offspring distances within plot')

wplotmofftx = {}
for plot in wplotmoff:
    tx = plot[:2]
    if not tx in wplotmofftx:
        wplotmofftx[tx] = []
    for samp in wplotmoff[plot]:
        wplotmofftx[tx].append(wplotmoff[plot][samp])
stats.kruskal(wplotmofftx['uc'],wplotmofftx['bc'],wplotmofftx['un'])

for i,txi in enumerate(wplotmofftx):
    for j,txj in enumerate(wplotmofftx):
        if i>j:
            print txi,txj
            js = [x for x in wplotmofftx[txj] if x != 'NA']
            Is = [x for x in wplotmofftx[txi] if x != 'NA']
            print np.median(Is),np.median(js)
            print stats.kruskal(Is,js)

for i,txi in enumerate(wplotmofftx):
    for j,txj in enumerate(wplotmofftx):
        if i>j:
            print txi,txj
            js = [x for x in wplotmofftx[txj] if x != 'NA']
            Is = [x for x in wplotmofftx[txi] if x != 'NA']
            print np.median(Is),np.median(js)
            print stats.kruskal(Is,js).pvalue

f = plt.figure()
vals = {}
for tx in txmf:
    vals[tx] = [val for val in wplotmofftx[tx] if not val == 'NA']
    print tx,np.mean(vals[tx]),np.median(vals[tx])
plt.boxplot([vals[tx] for tx in vals])
f.set_size_inches(6, 8)
plt.xticks(range(1,4,1),list([tx.upper() for tx in vals]))
plt.ylabel('Mother-Offspring Distances (m)')
f.savefig('/home/lindb/teakettle/postmoran/figs/wplotmofftx.pdf')

vals = {}
for tx in wplotmofftx:
    v = [val for val in wplotmofftx[tx] if not val == 'NA']
    vals[tx] = np.median(v)
    print tx,np.mean(v),np.median(v)

bar = plt.bar(range(len(vals)),list(vals.values()),color='grey')
for i,rect in enumerate(bar):
    h = rect.get_height()
    plt.text(rect.get_x() + rect.get_width()/2.0,h,['A','B','C'][i],ha='center',va='bottom')
plt.xticks(range(len(vals)), list([v.upper() for v in vals.keys()]))
plt.title('Mother-offspring distances within plot by treatment')

#### father-offspring distances

wplotdoff = {}
for plot in plotsamps:
    wplotdoff[plot] = {}
    for samp in plotseeds[plot]:
        wplotdoff[plot][samp] = wtd_dist(samp,'dad',byplot=True)

vals = {}
for plot in plotsamps:
    v = [val for val in wplotdoff[plot].values() if not val == 'NA']
    vals[plot] = np.mean(v)
    print plot,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Father-offspring distances within plot')

wplotdofftx = {}
for plot in wplotdoff:
    tx = plot[:2]
    if not tx in wplotdofftx:
        wplotdofftx[tx] = []
    for samp in wplotdoff[plot]:
        wplotdofftx[tx].append(wplotdoff[plot][samp])
stats.kruskal(wplotdofftx['uc'],wplotdofftx['bc'],wplotdofftx['un'])

for i,txi in enumerate(wplotdofftx):
    for j,txj in enumerate(wplotdofftx):
        if i>j:
            print txi,txj
            js = [x for x in wplotdofftx[txj] if x != 'NA']
            Is = [x for x in wplotdofftx[txi] if x != 'NA']
            print np.median(Is),np.median(js)
            print stats.kruskal(Is,js)

vals = {}
for tx in wplotdofftx:
    v = [val for val in wplotdofftx[tx] if not val == 'NA']
    vals[tx] = np.median(v)
    print tx,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Father-offspring distances within plot by treatment')

#### mother-father distances

# number
nmf = {}
nmftx = {}
for plot in plotseeds:
    nmf[plot] = []
    tx = plot[:2]
    if not tx in nmftx:
        nmftx[tx] = []
    for samp in plotseeds[plot]:
        dad = assignment[samp][0]
        dadplot = dad[:3]
        dadtx = dad[:2]
        if dadplot == plot:            
            nmf[plot].append(1)
            nmftx[tx].append(1)
        else:
            nmf[plot].append(0)
            nmftx[tx].append(0)
print [(plot,np.mean(nmf[plot])) for plot in nmf]
print [np.mean(nmftx[tx]) for tx in nmftx]

[(plot,sum(nmf[plot])) for plot in nmf]

vals = {}
for plot in plotsamps:
    v = [val for val in wplotmf[plot].values() if not val == 'NA']
    vals[plot] = np.mean(v)
    print plot,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-Father distances within plot')

wplotmftx = {}
for plot in wplotdoff:
    tx = plot[:2]
    if not tx in wplotmftx:
        wplotmftx[tx] = []
    for samp in wplotdoff[plot]:
        wplotmftx[tx].append(wplotdoff[plot][samp])
stats.kruskal(wplotmftx['uc'],wplotmftx['bc'],wplotmftx['un'])

for i,txi in enumerate(wplotmftx):
    for j,txj in enumerate(wplotmftx):
        if i>j:
            print txi,txj
            js = [x for x in wplotmftx[txj] if x != 'NA']
            Is = [x for x in wplotmftx[txi] if x != 'NA']
            print np.median(Is),np.median(js)
            print stats.kruskal(Is,js)

f = plt.figure()
vals = {}
for tx in txmf:
    vals[tx] = [val for val in wplotmftx[tx] if not val == 'NA']
    print tx,np.mean(vals[tx]),np.median(vals[tx])
plt.boxplot([vals[tx] for tx in vals])
f.set_size_inches(6, 8)
plt.xticks(range(1,4,1),list([tx.upper() for tx in vals]))
plt.ylabel('Father-Mother Distances (m)')
f.savefig('/home/lindb/teakettle/postmoran/figs/wplotmftx.pdf')

vals = {}
for tx in wplotmftx:
    v = [val for val in wplotmftx[tx] if not val == 'NA']
    vals[tx] = np.median(v)
    print tx,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-Father distances within plot by treatment')

# mean dispersal by plot to out-of-plot

#### mother-offspring distances

oplotmoff = {}
for plot in plotsamps:
    oplotmoff[plot] = {}
    for samp in plotseeds[plot]:
        oplotmoff[plot][samp] = wtd_dist(samp,'mom',byplot=True,without=True)

vals = {}
for plot in plotsamps:
    v = [val for val in oplotmoff[plot].values() if not val == 'NA']
    vals[plot] = np.mean(v)
    print plot,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Mother-offspring distances outside of plot')

#### father-offspring distances

oplotdoff = {}
for plot in plotsamps:
    oplotdoff[plot] = {}
    for samp in plotseeds[plot]:
        oplotdoff[plot][samp] = wtd_dist(samp,'dad',byplot=True,without=True)

vals = {}
for plot in plotsamps:
    v = [val for val in oplotdoff[plot].values() if not val == 'NA']
    vals[plot] = np.mean(v)
    print plot,np.mean(v),np.median(v)

plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)), list(vals.keys()))
plt.title('Father-offspring distances outside of plot')

# number of dispersals within plot

#### mother-offspring numbers

# number
nmoff = {}
nmofftx = {}
for plot in plotseeds:
    nmoff[plot] = []
    tx = plot[:2]
    if not tx in nmofftx:
        nmofftx[tx] = []
    for samp in plotseeds[plot]:
        mom = assignment[samp][0]
        momplot = mom[:3]
        momtx = mom[:2]
        if momplot == plot:            
            nmoff[plot].append(1)
            nmofftx[tx].append(1)
        else:
            nmoff[plot].append(0)
            nmofftx[tx].append(0)
print [(plot,np.mean(nmoff[plot])) for plot in nmoff]
print [np.mean(nmofftx[tx]) for tx in nmofftx]

for plot in plots:
    print plot, sum(nmoff[plot])

sum(nmoff[plot])

range(1,10,1)

# by plot
plt.bar(range(1,10,1),list([sum(nmoff[plot]) for plot in nmoff]))
plt.xticks(range(1,10,1),list(plots))
plt.title('Mother-offspring numbers within plot')

plt.bar(range(1,4,1),list([sum(nmofftx[tx]) for tx in nmofftx]))
plt.xticks(range(1,4,1),list(nmofftx.keys()))
plt.title('Mother-offspring numbers within tx')

# by tx
plt.bar(range(1,4,1),list([np.mean(nmofftx[tx]) for tx in nmofftx]))
plt.xticks(range(1,4,1),list([tx.upper() for tx in nmofftx.keys()]))
plt.title('Mother-offspring numbers within plot')

# scaled by number of seeds
vals = {}
for plot in plotsamps:
    lst = []
    for samp in nmoff[plot]:
        lst.append(nmoff[plot][samp])
    vals[plot] = np.mean(lst)/len(plotseeds[plot])
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Mother-offspring numbers within plot scaled by nseeds/plot')

# by tx scaled by nseeds
vals = {}
for tx in nmofftx:
    lst = []
    for samp in nmofftx[tx]:
        lst.append(nmofftx[tx][samp])
    vals[tx] = np.mean(lst)/len(txseeds[tx])
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Mother-offspring numbers within plot by tx scaled by nseeds/plot')

vals['bc']

np.mean(nmofftx['bc'])

np.mean(nmofftx['bc'])/len(plotseeds)

# scaled by number of adults
vals = {}
for plot in plotsamps:
    lst = []
    for samp in nmoff[plot]:
        lst.append(nmoff[plot][samp])
    vals[plot] = np.mean(lst)/len(plotadults[plot])
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Mother-offspring numbers within plot scaled by nadults/plot')

# scaled by possible parent-offspring pairs
vals = {}
for plot in plotsamps:
    lst = []
    for samp in nmoff[plot]:
        lst.append(nmoff[plot][samp])
    vals[plot] = np.mean(lst)/(len(plotadults[plot])*len(plotseeds[plot]))
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Mother-offspring numbers within plot scaled by nadults/plot')

#### father-offspring numbers

ndoff = {}
for plot in plotseeds:
    ndoff[plot] = []
    for samp in plotseeds[plot]:
        ndoff[plot].append(wtd_dist(samp,'dad',byplot=True,bynum=True))

vals = {}
for plot in plotsamps:
    lst = []
    for samp in ndoff[plot]:
        lst.append(ndoff[plot][samp])
    vals[plot] = np.mean(lst)
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Father-offspring numbers within plot')

vals = {}
for plot in plotsamps:
    lst = []
    for samp in ndoff[plot]:
        lst.append(ndoff[plot][samp])
    vals[plot] = np.mean(lst)/len(plotseeds[plot])
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Father-offspring numbers within plot scaled by nseeds/plot')

vals = {}
for plot in plotsamps:
    lst = []
    for samp in ndoff[plot]:
        lst.append(ndoff[plot][samp])
    vals[plot] = np.mean(lst)/len(plotadults[plot])
plt.bar(range(len(vals)),list(vals.values()))[-1]
plt.xticks(range(len(vals)),list(vals.keys()))
plt.title('Father-offspring numbers within plot scaled by nadults/plot')

#### mother-father numbers

# number
nmf = {}
nmftx = {}
for plot in plotseeds:
    nmf[plot] = []
    tx = plot[:2]
    if not tx in nmftx:
        nmftx[tx] = []
    for samp in plotseeds[plot]:
        mom = assignment[samp][0]
        momplot = mom[:3]
        momtx   = mom[:2]
        dad = assignment[samp][1]
        dadplot = dad[:3]
        dadtx = dad[:2]
        if dadplot == momplot:          
            nmf[plot].append(1)
        else:
            nmf[plot].append(0)
        if dadtx == momtx:
            nmftx[tx].append(1)
        else:
            nmftx[tx].append(0)
print [(plot,np.mean(nmf[plot])) for plot in nmf]
print [np.mean(nmftx[tx]) for tx in nmftx]

# by plot
plt.bar(range(1,10,1),list([sum(nmf[plot]) for plot in nmf]))
plt.xticks(range(1,10,1),list(plots))
plt.title('Father-mother numbers within plot')

plt.bar(range(1,4,1),list([sum(nmftx[tx]) for tx in nmftx]))
plt.xticks(range(1,4,1),list(nmftx.keys()))
plt.title('Father-mother numbers within tx')

# by tx
plt.bar(range(1,4,1),list([np.mean(nmftx[tx]) for tx in nmftx]))
plt.xticks(range(1,4,1),list([tx.upper() for tx in nmftx.keys()]))
plt.title('Father-mother numbers within plot')

# number of dispersals outside of plot

#### mother-offspring numbers

plots = ['bc1','bc2','bc3','uc1','uc2','uc3','un1','un2','un3']

# for each plot, which known mothers dispersed seeds to known seedlings out of plot
momout = {}
for plot in plots:
    if not plot in momout:
        momout[plot] = 0
    for seed in plotseeds[plot]:
        mom = assignment[seed][0] # the most likely mother
        if mom != 'out-of-plot':  # if a known mother
            print seed, mom
            momplot = mom[:3]
            if not momplot in momout:
                momout[momplot] = 0
            if momplot != plot:   # if a known mother from another plot dispersed seed to this plot
                momout[momplot] += 1 # count it
                print 'counting'
        else: # count 
momout # there are no known mothers that dispersed seed to another plot

# num of seedlings with out-of-plot mother, don't need to add in-plot mother dispersing to 
                                                    # known seedling outside of plot bc never happens
outdisp = {}
for plot in plots:
    tx = plot[:2]
    outdisp[plot] = 0
    for samp in plotseeds[plot]:
        n = 0
        mom = assignment[samp][0]
        momtx = mom[:2]
        if mom == 'out-of-plot': # if un-IDed
            n += 1
        elif momtx != tx:        # or from another plot
            n += 1
            print 'mama out'     # never prints
        outdisp[plot] += n
outdisp

# plot out-of-plot events
f = plt.figure()
outdisptx = {}
for plot in outdisp:
    tx = plot[:2]
    if not tx in outdisptx:
        outdisptx[tx] = []
    outdisptx[tx].append(outdisp[plot])
yerr = [np.std(outdisptx[tx]) for tx in outdisptx]
plt.bar(range(1,4,1),list([np.mean(outdisptx[tx]) for tx in outdisptx]),yerr=yerr,color='gray')
plt.xticks(range(1,4,1),list([tx.upper() for tx in outdisptx]))
plt.ylabel('Mother-offspring out-of-plot dispersal events')
plt.ylim(0,170)
f.savefig('/home/lindb/teakettle/postmoran/figs/moff_out-of-plot_disp_events.pdf')

print stats.kruskal(outdisptx['bc'],outdisptx['uc'],outdisptx['un'])

# ratio of within to without of plot
nmoffsum = {}
outtoin = {}
for plot in nmoff:
    nmoffsum[plot] = sum(nmoff[plot])
    outtoin[plot] = nmoffsum[plot] / outdisp[plot]
    print plot, outtoin[plot]

# plot num in-plot dispersal
f = plt.figure()
nmoffbytx = {}
for plot in nmoffsum:
    tx = plot[:2]
    if not tx in nmoffbytx:
        nmoffbytx[tx] = []
    nmoffbytx[tx].append(nmoffsum[plot])
nmoffbytx
yerr = [np.std(nmoffbytx[tx]) for tx in nmoffbytx]
plt.bar(range(1,4,1),list([np.mean(nmoffbytx[tx]) for tx in nmoffbytx]),yerr=yerr,color='gray')
plt.xticks(range(1,4,1),list([tx.upper() for tx in nmoffbytx]))
plt.ylabel('Mother-offspring in-plot dispersal events')
plt.ylim((0,90))
f.savefig('/home/lindb/teakettle/postmoran/figs/moff_in-plot_disp_events.pdf')

print stats.kruskal(nmoffbytx['bc'],nmoffbytx['uc'],nmoffbytx['un'])

# plot ratio
plt.bar(range(1,10,1),list(outtoin.values()))
plt.xticks(range(1,10,1),list(outtoin.keys()))
plt.title('In-to-out ratio mother-offspring numbers')

range(1,4,1)

# combine to treatment
rattx = {}
for tx in nmofftx:
    rattx[tx] = []
    for i in [1,2,3]:
        plot = "%s%s" % (tx,i)
        rattx[tx].append(outtoin[plot])
rattx

# plot mean ratio across replicates by tx
f = plt.figure()
yerr = [np.std(rattx[tx]) for tx in rattx]
plt.bar(range(1,4,1),list([np.mean(rattx[tx]) for tx in rattx]),yerr=yerr,color='gray')
plt.xticks(range(1,4,1),list([k.upper() for k in rattx.keys()]))
#plt.title('In-to-out ratio mother-offspring numbers by tx')
plt.ylim((0,0.7))
plt.ylabel('Mother-offspring dispersal events ratio')
f.savefig('/home/lindb/teakettle/postmoran/figs/ratio_in-to-out_by_tx.pdf')

print stats.kruskal(rattx['bc'],rattx['uc'],rattx['un'])

#### father-mother numbers

plots

txs = ['un','uc','bc']
txs

# get pedigree info and combine by plot
plotpars = {}
inplots = {}
for plot in plots:
    plotpars[plot] = []
    inplots[plot] = []
for samp in assignment:
    outs = 0
    mom,dad = assignment[samp]
    if not mom == 'out-of-plot':
        momplot = mom[:3]
        momtx = mom[:2]
    else:
        outs += 1
    if not dad == 'out-of-plot':
        dadplot = dad[:3]
        dadtx = dad[:2]
    else:
        outs += 1
        
    if outs == 0 : # if both parents sampled
        if dadplot == momplot: # 
            inplots[dadplot].append(1)
        else: # each plot gets a tick
            plotpars[dadplot].append(1)
            plotpars[momplot].append(1)
    if outs == 1 : # if only one parent sampled
        if not mom == 'out-of-plot': # if the mom is sampled but not the dad
            plotpars[momplot].append(1)
        if not dad == 'out-of-plot': # if the dad is sampled but not the mom
            plotpars[dadplot].append(1)
    if outs == 2 : # if both parents unsampled
        1 + 1 # can't bin these by plots 

from collections import Counter

# for each seed (for each mother)
# count if received pollen from out-of-plot
c = Counter()
for samp in assignment:
    mom,dad = assignment[samp]
    if not (mom,dad) == ('out-of-plot','out-of-plot'): # if at least one of them is sampled
        if not mom == 'out-of-plot':            
            momplot = mom[:3]
            dadplot = dad[:3]
            if momplot != dadplot: # dad could be from another plot or 'out' ('out-of-plot'[:3])
                c[momplot] += 1
c    

# count in-plot similarly
d = Counter()
for samp in assignment:
    mom,dad = assignment[samp]
    if mom != 'out-of-plot' and dad != 'out-of-plot': # if both parents sampled
        momplot = mom[:3]
        dadplot = dad[:3]
        if momplot == dadplot:
            d[momplot] += 1
d

# plot in-plot events
f = plt.figure()
dsums = {}
for plot in d:
    tx = plot[:2]
    if not tx in dsums:
        dsums[tx] = []
    dsums[tx].append(d[plot])
yerr = [np.std(dsums[tx]) for tx in dsums]
plt.bar(range(1,4,1),list([np.mean(dsums[tx]) for tx in dsums]),yerr=yerr,color='gray')
plt.xticks(range(1,4,1),list([tx.upper() for tx in dsums]))
plt.ylabel('Father-mother in-plot dispersal events')
f.savefig('/home/lindb/teakettle/postmoran/figs/mf_in-plot_disp_events.pdf')

print stats.kruskal(dsums['bc'],dsums['uc'],dsums['un'])

# plot in-plot events
f = plt.figure()
csums = {}
for plot in c:
    tx = plot[:2]
    if not tx in csums:
        csums[tx] = []
    csums[tx].append(c[plot])
yerr = [np.std(csums[tx]) for tx in csums]
plt.bar(range(1,4,1),list([np.mean(csums[tx]) for tx in csums]),yerr=yerr,color='gray')
plt.xticks(range(1,4,1),list([tx.upper() for tx in csums]))
plt.ylabel('Father-mother out-of-plot dispersal events')
f.savefig('/home/lindb/teakettle/postmoran/figs/mf_out-of-plot_disp_events.pdf')

csums['bc'].append(0)

print stats.kruskal(csums['bc'],csums['uc'],csums['un'])

# plot ratio
f = plt.figure()
rat = {}
for plot in c:
    tx = plot[:2]
    if not tx in rat:
        rat[tx] = []
    rat[tx].append(d[plot]/c[plot])
yerr = [np.std(rat[tx]) for tx in rat]
plt.bar(range(1,4,1),list([np.mean(rat[tx]) for tx in rat]),yerr=yerr,color='gray')
plt.xticks(range(1,4,1),list([tx.upper() for tx in rat]))
plt.ylabel('Father-mother dispersal events ratio')
f.savefig('/home/lindb/teakettle/postmoran/figs/mf_ratio_in-to-out_by_tx.pdf')

print stats.kruskal(rat['bc'],rat['uc'],rat['un'])

[(tx,np.mean(rat[tx])) for tx in rat]

d

#### mother-father numbers

# what I think I can do
1. father-mother distance greater/smaller than expected by chance
    1. null would be distribution of all in-plot adults to any adults
1. seedling-mother distance greater/smaller than expected by chance
    1. null would be distribution of all seedling-mother distances
1. father-seedling distances "" ""
    1. "" ""

# pollen pool heterogeneity

# get a list of seedlings per mom with potential fathers
moms = {}
for seed in assignment:
    mom,dad = assignment[seed]
    if not mom == 'out-of-plot':
        if not dad == 'out-of-plot':
            if not mom in moms:
                moms[mom] = {}
                moms[mom]['seeds'] = []
                moms[mom]['dads'] = []
            moms[mom]['seeds'].append(seed)
            moms[mom]['dads'].append(dad)

# include only those with greater than one seed
nums = []
for mom in moms:
    l = len(moms[mom]['seeds'])
    if l > 1:
        print mom, l
        nums.append(l)

plt.hist(nums)[-1]
